requirements.txt
fastapi
uvicorn
line-bot-sdk
openai
chromadb
langchain
langgraph
python-dotenv